add the pr list command
This adds the ability to list the open pull requests on a single repo.

When there are more than 20 open pull requests it grabs the newest ten and oldest ten, while lightly insulting you...

This also converts the 'repo_match()' method to a mixin method as it is used multiple places.