[v0.0.7] make default_team_slug not a required option
Previously the `config.handlers.github.default_team_slug` option was required. Otherwise, if you didn't specify a team weird things may have happened? I never actually tried it...

This patch updates this logic so that if no team slug or an invalid one is provided in the command, and the default team slug is not set, then it should not specify a :team_id for the operation. For repo creation, this means only organization owners can access the repo which seems like a sane default.

In addition, one of the signatures of a private method in the GithubRepo handler was changed:

`team_by_slig(slug, organization)` => `team_id_by_slug(slug, organization)`