GitHub team management v1
This adds the first version of the team management to `lita-github` with the following functionality:

* the ability to list the teams of on repo
* the ability to add teams to the repo
* the ability to remove teams from the repo
* the ability to list all teams in your organization

The command to add and remove teams is disabled by default in the config.

The commands are as follows, with some aliases being supported as well.

* `!gh repo teams PagerDuty/lita-github`
  * show the teams on a repo, sorted alphabetically
* `!gh repo team rm <TEAM_ID|TEAM_SLUG> PagerDuty/lita-gitub`
  * add a team to a repo, can either use their team ID or their slug. Requires confirmation.
* `!gh repo team rm <TEAM_ID|TEAM_STUB> PagerDuty/lita-gitub`
  * remove a team to a repo, can either use their team ID or their slug. Requires confirmation.
* `!gh org teams PagerDuty` or `!gh teams PagerDuty`
  * list all the teams of your organization, this uses the default Org so it can be omitted

To enable in the config:

```Ruby
config.handlers.github.repo_team_add_enabled = true
config.handlers.github.repo_team_rm_enabled  = true
```