delete old OPT_REGEX and replace with E_OPT_REGEX
This replaces the `OPT_REGEX` constant with the one that was `E_OPT_REGEX`. The `E_OPT_REGEX` regex was backwards compatible, and extended the option ability to be able to use spaces in your option.

All references to `E_OPT_REGEX` have been updated to point to `OPT_REGEX`.