Increment to v0.2.1
This release fixes an issue introduced in v0.2.0 where the configuration
option allowing multiple teams to be added to a newly created repository
by default caused no teams to be added at all.
The original functionality has now been restored, and the option to add
multiple teams to a newly created repo is working as expected.